Adding auth folder to handle authentication related use cases. Refactoring level 101
