+++ apps/endatix-hub/features/auth/application/actions/login.action.ts
@@ -1,4 +1,4 @@
-"use server";
+'use server'
 
 import { AuthService } from "@/lib/auth-service";
 import {
