+++ apps/endatix-hub/features/auth/application/actions/logout.action.ts
@@ -1,4 +1,4 @@
-"use server";
+'use server'
 
 import { AuthService } from "@/lib/auth-service";
 import { redirect } from 'next/navigation'
