+++ apps/endatix-hub/features/auth/ui/login-form.tsx
@@ -5,7 +5,7 @@ import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import Link from "next/link";
 import { useActionState } from "react";
-import { loginAction } from "../login.action";
+import { loginAction } from "../application/actions/login.action";
 import { showComingSoonMessage } from "@/components/layout-ui/teasers/coming-soon-link";
 import { Spinner } from "@/components/loaders/spinner";
